% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotfunsample.R
\docType{data}
\name{defaultoptions.funsample}
\alias{defaultoptions.funsample}
\title{Plotting defaults for funsample
List of defaults for plotting \code{\link{funsample}}-objects}
\format{\preformatted{List of 6
 $ xlab: chr "t"
 $ ylab: chr "X(t)"
 $ main: chr ""
 $ col : NULL
 $ lwd : NULL
 $ lty : NULL
 - attr(*, "class")= chr [1:2] "simplist" "list"
}}
\usage{
defaultoptions.funsample
}
\description{
Plotting defaults for funsample
List of defaults for plotting \code{\link{funsample}}-objects
}
\keyword{internal}

