% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funsample.R
\name{funsample}
\alias{funsample}
\alias{is.funsample}
\title{Class for a sample of functions}
\usage{
funsample(funs, ..., arglim = c(0, 1))

is.funsample(x)
}
\arguments{
\item{funs}{a single \code{function} taking one argument,
or a \code{list} of objects of class'\code{function}, that also take one argument.}

\item{...}{predefined plot options, may be given as \code{\link{simplist}}s.}

\item{arglim}{numeric vector, an interval the functions in \code{funlist} are defined on,
, defaults to \code{c(0, 1)}.}

\item{x}{an \code{R} object to be checked.}
}
\value{
\code{funsample} returns an object of class \code{funsample}, that is a function of one variable
adorned with attributes
\tabular{ll}{
\code{funs}  \tab{a list of \code{function}s}
\cr\code{arglim} \tab{numeric, allowed range of function arguments}
\cr\code{groupsize} \tab{integer, length of \code{funs}}
\cr\code{options} \tab \code{\link{simplist}} of plot options, see \code{{plot.funsample}}
}
}
\description{
An object of class \code{funsample} can be considered both as a multivariate
function of one argument and as a list of univariate functions. It is generated by function \code{funsample}.

\code{is.funsample} returns \code{TRUE} if \code{x} is an object
of class \code{funsample}
}
\details{
When called with an argument \code{x}, an object of class \code{funsample} applies
all functions contained in \code{fun} to \code{x}, using R-function \code{\link{sapply}}.
The argument limit \code{arglim} does not matter, i.e. also values \code{x} outside
the "allowed range" are processed.

The return value of a \code{funsample} is a matrix. It has \code{\link{colnames}}
if the list \code{funs} was named.

Currently, only functions of one variable are allowed, i.e., for any element
\code{foo} of \code{funlist}, \code{foo(x)} needs to return a numeric value as
long as \code{x} is within \code{arglim}. The intervall \code{arglim} is necessary
for plotting and for coercing into an \code{\link{fdsample}} object.

Predefining plot options may be particularly useful for plot annotation,
such as i.e. \code{xlab}, \code{ylab} and \code{main}. They may be also given
as a \code{plutils::\link{simplist}}. Options with same name are overriden
from left to right, i.e., the last one given counts.
}
\examples{
myfuns <- funsample(list(sin = sin, cos = cos), arglim = c(0, 2*pi))
myfuns(c(0, pi/4, pi/2))
}
\author{
Ute Hahn \email{ute@imf.au.dk}
}

