\name{is.fdsample}
\alias{is.fdsample}
\title{Check whether object is a function sample}
\usage{
  is.fdsample(x)
}
\arguments{
  \item{x}{any R-object}
}
\value{
  \code{TRUE}, if \code{x} belongs to class
  \code{fdsample}, otherwise \code{FALSE}
}
\description{
  Checks whether object is of type \code{fdsample}.
}

