% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rankcounttest.R
\name{rankCount.test}
\alias{rankCount.test}
\title{Rank count test of goodness of fit}
\usage{
rankCount.test(obs, sim, alternative = c("two.sided", "less", "greater"),
  inclprob = c(0.95))
}
\arguments{
\item{obs}{object of class \code{fdsample}, the observed curve.}

\item{sim}{object of class \code{fdsample}, the group of curves to which \code{obs} is
compared.}

\item{alternative}{a character string specifying the alternative hypothesis, one of
\code{"two.sided"} (default), \code{"less"} or \code{"greater"}. May be abbreviated.}

\item{inclprob}{a numerical vector of inclusion probabilities
of the envelopes to be plotted, for use in \code{\link{plot.envtest}}}
}
\description{
Test if an observed curve matches a sample of simulated curves (or
a group of other observations). Refines the result of the rank envelope test, but
does not have a direct graphical representation.
}
\details{
The observed curve, represented by the \code{\link{fdsample}} object
\code{obs} is compared to simulated curves collected
in the \code{fdsample} object '\code{sim}.
The two sets of curves have to share the same argument values, and \code{obs} is
supposed to contain only one curve.

\code{alternative == "less"}
is the one-sided alternative meaning that the observed curve
has (some) smaller function values than the simulated curves, and
\code{alternative == "greater"} is the opposite one-sided alternative.

The p-value is obtained by ranking the curves according to the number of
minimum pointwise ranks obtained in any point of the curve (curves are actually
represented as vectors), see Myllymaki et al. (2015).
}
\examples{
# make a sample of sinus curves
tt <- seq(0, 2*pi, length = 20)
sinsim <- replicate(5000, sin(tt) + cumsum(rnorm(20, 0, 0.01)))
sinobs <- sin(tt - pi/50) + cumsum(rnorm(20, 0, 0.01))
sim <- fdsample(tt, sinsim)
obs <- fdsample(tt, sinobs)

testresult <- rankCount.test(obs, sim)
print(testresult)
}
\author{
Ute Hahn, \email{ute@imf.au.dk}
}
\references{
M. Myllymaki, T. Mrkvicka, P. Grabarnik, H. Seijo and U.Hahn (2015)
\emph{Global envelope tests for spatial processes},
\url{http://arxiv.org/abs/1307.0239v3}.
}

