% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/envelopetest.R
\name{rankEnv.test}
\alias{rankEnv.test}
\title{Envelope test of goodness of fit}
\usage{
rankEnv.test(obs, sim, alternative = c("two.sided", "less", "greater"),
  inclprob = 0.95, includeobs = TRUE)
}
\arguments{
\item{obs}{object of class \code{fdsample}, the observed curve.}

\item{sim}{object of class \code{fdsample}, the group of curves to which \code{obs} is
compared.}

\item{alternative}{a character string specifying the alternative hypothesis, one of
\code{"two.sided"} (default), \code{"less"} or \code{"greater"}. May be abbreviated.}

\item{inclprob}{a numerical vector of inclusion probabilities
of the envelopes to be plotted, for use in \code{\link{plot.envtest}}}

\item{includeobs}{logical, if TRUE, observed curve is also used in the envelope}
}
\description{
Test if an observed curve matches a sample of simulated curves (or
a group of other observations). A global envelope test is performed, with
p-value corresponding to the most extreme pointwise percentile the observed
curve reaches among all curves.
}
\details{
The observed curve, represented by the \code{\link{fdsample}} object
\code{obs} is compared to simulated curves collected
in the \code{fdsample} object '\code{sim}.
The two sets of curves have to share the same argument values, and \code{obs} is
supposed to contain only one curve.

\code{alternative == "less"}
is the one-sided alternative meaning that the observed curve
has (some) smaller function values than the simulated curves, and
\code{alternative == "greater"} is the opposite one-sided alternative.

The test corresponds to the rank envelope test by Myllymaki et. al (2013, 2015),
and to the procedure described in Davison and Hinkley (1997), Equation (4.17).
The  p-value is obtained by ranking the curves according to the minimum pointwise
rank obtained in any point of the curve -- note that the curves are actually
represented as vectors.

The result of the test can be plotted, see \code{\link{plot.envtest}}.
}
\examples{
# make a sample of sinus curves
tt <- seq(0, 2*pi, length = 20)
sinsim <- replicate(5000, sin(tt) + cumsum(rnorm(20, 0, 0.01)))
sinobs <- sin(tt - pi/50) + cumsum(rnorm(20, 0, 0.01))
sim <- fdsample(tt, sinsim)
obs <- fdsample(tt, sinobs)

testresult <- rankEnv.test(obs, sim)
print(testresult)
plot(testresult)
}
\author{
Ute Hahn, \email{ute@imf.au.dk}
}
\references{
M. Myllymaki, T. Mrkvicka, H. Seijo and P. Grabarnik (2013)
\emph{Global envelope tests for spatial processes},
\url{http://arxiv.org/abs/1307.0239v2}.

M. Myllymaki, T. Mrkvicka, P. Grabarnik, H. Seijo and Ute Hahn (2015)
\emph{Global envelope tests for spatial processes},
\url{http://arxiv.org/abs/1307.0239v3}.

Davison, A.C. and Hinkley, D.V. (1997) \emph{Bootstrap Methods and their
Applications}, Cambridge University Press, Cambridge.
}

