% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ploturfunction.R
\name{legendinfo.urfunction}
\alias{legendinfo.urfunction}
\title{Legend information for an urfunction}
\usage{
legendinfo.urfunction(urfu)
}
\description{
Legend information for an urfunction
}
\keyword{internal}

