\name{yrange}
\alias{yrange}
\title{range of function values}
\usage{
  yrange(x, includy = NULL)
}
\arguments{
  \item{x}{the data to be inspected}

  \item{includy}{anything to be included in the range}
}
\value{
  Numeric vector of length two.
}
\description{
  Range of \code{fvals} contained in an \code{fdasample}
  object, purpose: convenient way to find plot y-limits.
}

