% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fdsample.R
\name{fdsample}
\alias{fdsample}
\title{Make a functional data sample}
\usage{
fdsample(args, fvals, ...)
}
\arguments{
\item{args}{numeric, array, the function arguments, dimension dimarg}

\item{fvals}{numeric, array of function values, dimension: dimarg x groupsize}

\item{...}{predefined plot options or \code{\link{simplist}}s, see the details
and \code{\link{plot.fdsample}}}
}
\value{
An object of class \code{fdsample}, that is a list with elements
\tabular{ll}{
\code{args} \tab{the function arguments}
\cr\code{fvals}  \tab{the function value samples}
\cr\code{dimarg} \tab{integer, dimension of \code{x}}
\cr\code{groupsize} \tab{integer, number of functions y included in the list}
\cr\code{options} \tab list of plot options, see \code{\link{plot.fdsample}}
}
}
\description{
Generate an object of class \code{fdsample}.
}
\details{
Predefining plot options may be particularly useful for plot
annotation, i.e. \code{xlab}, \code{ylab} and \code{main}. They may be also given
as a \pkg{plutils}-\code{\link{simplist}}. Options with same name are overriden
from left to right, i.e., the last one given counts.
}

