\docType{data}
\name{defaultoptions.funsample}
\alias{defaultoptions.funsample}
\title{Plotting defaults for funsample
List of defaults for plotting \code{\link{funsample}}-objects}
\format{List of 6
 $ xlab: chr "t"
 $ ylab: chr "X(t)"
 $ main: chr ""
 $ col : NULL
 $ lwd : NULL
 $ lty : NULL
 - attr(*, "class")= chr [1:2] "simplist" "list"}
\usage{
  defaultoptions.funsample
}
\description{
  Plotting defaults for funsample List of defaults for
  plotting \code{\link{funsample}}-objects
}
\keyword{datasets}
\keyword{internal}

