% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/envelopetest.R
\name{plot.envtest}
\alias{plot.envtest}
\title{Plot the result of an envelope test}
\usage{
\method{plot}{envtest}(x, ..., col.obs = "red")
}
\arguments{
\item{x}{test result to be plotted, an object of type \code{envtest}}

\item{...}{arguments passed to plot methods}

\item{col.obs}{color for plotting the observations}
}
\description{
Plots simultaneous envelopes of a rank envelope test and
prints the exact inclusion probabilities.
}
\details{
Objects of type \code{envtest} are generated by \code{\link{rankEnv.test}}
which calculates the result of a test comparing an observed curve to a family
of other curves (most often simulated ones).

Method \code{plot.envtest} plots envelopes according to the inclusion probabilities
given while
calculating the test, and adds the observed curve. The most extreme points of
the observed curve, i.e. the points that are reponsible for the curve's p-value,
are marked by points.

For an example, see \code{\link{rankEnv.test}}
}

