\name{print.fdsample}
\alias{print.fdsample}
\title{Print brief details of an xy-list}
\usage{
  \method{print}{fdsample} (x, ...)
}
\arguments{
  \item{x}{object of type \code{\link{fdsample}}}

  \item{...}{ignored}
}
\description{
  Prints a very short summary of the \code{fdsample}-object
  in the argument
}

